requirements.txt
numpy>=1.20.0
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
matplotlib>=3.3.0
scipy>=1.7.0
tensorboard>=2.4.0
